+++ src/index.ts
@@ -228,7 +228,11 @@ export type { Slice } from "./types/value/slice";
 export type { SharedSlice } from "./types/value/sharedSlice";
 export type { SharedSliceVariation } from "./types/value/sharedSliceVariation";
 
-export type { FieldState, AnyRegularField } from "./types/value/types";
+export type {
+	FieldState,
+	AnyRegularField,
+	AnySlicePrimaryField,
+} from "./types/value/types";
 
 // Models - Types representing Prismic content models.
 export { CustomTypeModelFieldType } from "./types/model/types";
@@ -300,6 +304,7 @@ export type { CustomTypeModelSeparatorField } from "./types/model/separator";
 export type {
 	CustomTypeModelField,
 	CustomTypeModelFieldForGroup,
+	CustomTypeModelFieldForSlicePrimary,
 } from "./types/model/types";
 
 // API - Types representing Prismic Rest API V2 responses.